+++ src/LionWeb.Generator/GeneratorExtensions/Extensions.cs
@@ -17,7 +17,9 @@
 
 namespace LionWeb.Generator.GeneratorExtensions;
 
+using Core.M3;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
@@ -54,5 +56,65 @@ public static class Extensions
                 onPersistedFile?.Invoke(copy, file);
             }
         }
+    }
+
+    /// <inheritdoc cref="AddUsing(CompilationUnitSyntax, string)"/>
+    public static CompilationUnitSyntax AddUsing(this CompilationUnitSyntax compilationUnit, Type type) =>
+        compilationUnit.AddUsing(type.Namespace!);
+
+    /// <summary>
+    /// Adds "global::<i><paramref name="nameSpace"/></i>" to the <i>usings</i> of the first namespace in <paramref name="compilationUnit"/>.
+    /// </summary>
+    /// <returns>Modified version of <paramref name="compilationUnit"/>.</returns>
+    public static CompilationUnitSyntax AddUsing(this CompilationUnitSyntax compilationUnit, string nameSpace)
+    {
+        var ns = compilationUnit.Members.OfType<FileScopedNamespaceDeclarationSyntax>().First();
+        return compilationUnit.ReplaceNode(
+            ns,
+            ns.AddUsings(UsingDirective(ParseName($"global::{nameSpace}")))
+        );
+    }
+
+    /// <summary>
+    /// Adds <paramref name="annotations"/> to the initializers in the language constructor in <paramref name="compilationUnit"/>.
+    /// </summary>
+    /// <returns>Modified version of <paramref name="compilationUnit"/>.</returns>
+    public static CompilationUnitSyntax AddM2Annotations(this CompilationUnitSyntax compilationUnit, Correlator correlator,
+        Dictionary<IKeyed, CollectionExpressionSyntax> annotations)
+    {
+        var correlations = annotations.Keys
+            .SelectMany(correlator.FindAll<KeyedToLanguageConstructorCorrelation>);
+
+        foreach (var correlation in correlations)
+        {
+            var assignment = correlation.LookupIn(compilationUnit);
+            if (assignment.Right is ImplicitObjectCreationExpressionSyntax
+                {
+                    ArgumentList.Arguments:
+                    [
+                        {
+                            Expression: ParenthesizedLambdaExpressionSyntax
+                            {
+                                Body: ObjectCreationExpressionSyntax
+                                {
+                                    Initializer: {} init
+                                }
+                            }
+                        }
+                    ]
+                })
+            {
+                compilationUnit = compilationUnit.ReplaceNode(
+                    init,
+                    init.AddExpressions(AssignmentExpression(
+                        SyntaxKind.SimpleAssignmentExpression,
+                        IdentifierName("Annotations"),
+                        annotations[correlation.Keyed]
+                    ))
+                );
+            }
+        }
+
+        return compilationUnit;
     }
 }